Updating wgpu from 0.14.2 to 0.15.1
Updated all code necessary to run on wgpu 0.15. For now the new
view_formats aren't really used and just contain the same format as the
texture. The new instance descriptor uses the FXC compiler for Windows
and all others. For all others it shouldn't matter as it is Windows
only. For Windows the possibility of using the DXC compiler should be
explored as it seems to be a better alternative to FXC
(https://docs.rs/wgpu-types/0.15.1/wgpu_types/enum.Dx12Compiler.html#variants).