`Gpu` ownership cleanup in `gpu` and `rendering` crates
The primary goal of this PR is to improve the readability of functions which
interact with the graphics API. Previously, most GPU-related types cloned the
`Arc<Gpu>` and stored locally next to their actual data. Then the member
functions were able to access the `Gpu` via `self.gpu`. While this can save
some typing, it obscures the fact that a function might be doing something
noteworthy with the graphics API, like writing data into a GPU buffer.

We can achieve this goal by changing the "shared ownership of `Arc<Gpu>`" into
just a simple `&Gpu`. In practice this means removing almost all `Arc<Gpu>`
from structures, and instead passing `&Gpu` as a function argument down the
callstack. In most cases the `Gpu` already exists somewhere higher in the
callstack, so it doesn't need to be fetched again. A notable exception is all
the AssetKey-types where you must get the `Arc<Gpu>` from the asset cache.

The only three places where `Arc<Gpu>` still lives are:

- `AppResources`
- `World`
- `AssetCache`

This PR shouldn't cause any functional changes to the renderer.