feat: vite web assembly
Vite is easier to use and a lot faster than Webpack.

For instance; there is no multi second long "compiling modules" when the
bundler runs, but it is rather instant and HMR works as expected.

Additionally, including assets files with Vite is a breeze and there is
no need for CopyWebPack plugins and alike