fix: mixup between buffer item count and byte sizes and offsets
Additionally, this makes the type-safe version consistent with
conventional Rust containers such as slices and `Vec`s and reduces the
need for lossy integer conversions as conversions now mostly occur to
larger target integer types.

This is prominent for wasm where `usize: u32` and casting back and forth
between `u64` and `u32` is lossy.

Furthermore, this makes commit makes bounds checking stricter to ensure
that no overread and overwrite, as well as uninitialized reads or
incorrectly bound ranges occur,