feat(build): bring back OptiVorbis
Even though OptiVorbis was deemed appropriate for the sound build
pipeline of Ambient in #619, its inclusion was later reverted in #751
due to a significant impact in compile times for a transitive
dependency.

The good news is that, at long last, a new version of OptiVorbis is
here, with refactored internals that got rid of the troublesome
dependencies! Several other changes have also been made, which can be
read here: https://github.com/OptiVorbis/OptiVorbis/blob/master/CHANGELOG.md#020---2023-12-11

While at it, I've also updated the dependency on `vorbis_rs`. This new
version of `vorbis_rs` brings a new builder for instantiating encoders
as I initially discussed with @marceline-cramer in
https://github.com/ComunidadAylas/vorbis-rs/pull/9, and other minor
fixes and QoL changes you can read about in the changelog if interested.